feat: create function to mark order as delivered
The names of the arrays for the products were changed for more clarity: ordersReady changed to ordersDelivering,
array for orders delivered created. Functions to show those orders were created in waiter order status.
Cook orders ready updated accordingly, both html as ts.
Updated service for the feature, changing the adding to dateProcessed for the correct function.